+++ SpiderKeeper/app/util/__init__.py
@@ -1,0 +1,4 @@
+def project_path():
+    import inspect, os
+    this_file = inspect.getfile(inspect.currentframe())
+    return os.path.abspath(os.path.dirname(this_file)+'/../')