+++ CHANGELOG.md
@@ -1,11 +1,19 @@
 # SpiderKeeper Changelog
-##v0.2.0
+## 1.0.0 (2017-03-30)
+- refactor
+- support py3
+- optimized api
+- optimized scheduler
+- more scalable (can support access multiply spider service)
+- show running stats
+
+## 0.2.0 (2016-04-13)
 - support view job of multi daemons.
 - support run on multi daemons.
 - support choice running daemon automaticaly.
 
-##v0.1.1
+## 0.1.1 (2016-02-16)
 - add status monitor(https://github.com/afaqurk/linux-dash)
 
-##v0.1.0
+## 0.1.0 (2016-01-18)
 - initial.